requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
pydantic==2.9.0
python-dotenv==1.0.1
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
databricks-sql-connector==3.0.0
